+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+- Add `CallableTree::Node::Internal#reject` that returns a new node instance without rejected child nodes.
+- Add `CallableTree::Node::Internal#reject!` that make destructive change.
 
 ## [0.2.1] - 2021-07-24
 - Add `CallableTree::Node#root?`.